send money finished

+++ src/main/java/com/banking/services/checkingService.java
@@ -8,7 +8,7 @@ public interface checkingService {
     public checkingAccountInfo create (checkingAccountInfo checkingAccountInfo);
     List<checkingAccountInfo> findall();
     List<checkingAccountInfo> findallbyid(int id);
-    checkingAccountInfo get(int id);
+    checkingAccountInfo get(long id);
     checkingAccountInfo update(checkingAccountInfo checkingAccountInfo);
     void delete (int id);
 }
